Adds Movie Controller and moves logic of card & popup rendering there
